refactor: 프로젝트 폴더 구조 개선
- src 디렉토리를 기능별로 모듈화 (ai, config, core, utils)
- JSON 데이터 파일을 data 디렉토리로 정리하여 관리
- 환경 변수 파일 (.env, .env.example) 명확히 구분
- 백업 파일 관리 폴더 유지 (backup)
- 코드 가독성과 유지보수를 위한 파일 구조 개선
