+++ .env.example
@@ -1,0 +1,2 @@
+OEPNAI_API_KEY=YOUR-API-KEY
+GOOGLE_API_KEY=YOUR-API-KEY
