+++ src/core/json_handler.py
@@ -9,7 +9,7 @@ import json
 from datetime import datetime
 import shutil
 from typing import Dict, Any
-from src.ai_gemini import AIManager
+from ai.ai_gemini import AIManager
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
@@ -51,7 +51,7 @@ def update_sort_data(sort_data, sorted_data, src_path='./data/sort_data.json') -
         os.remove(old_backup)
 
     timestamp = datetime.now().strftime("%Y%m%d_%H%M")
-    backup_path = f"./backup/sort_data_{timestamp}.json"
+    backup_path = f"./data/backup/sort_data_{timestamp}.json"
 
     if os.path.exists(src_path):
         shutil.copy(src_path, backup_path)
@@ -113,7 +113,7 @@ def update_docu_data(docu_data, docued_data, src_path='./data/docu_data.json') -
     ai = AIManager()
 
     timestamp = datetime.now().strftime("%Y%m%d_%H%M")
-    backup_path = f"./backup/docu_data_{timestamp}.json"
+    backup_path = f"./data/backup/docu_data_{timestamp}.json"
 
     if os.path.exists(src_path):
         shutil.copy(src_path, backup_path)
